Moved connected components to containers folder
